third_party/WebKit/LayoutTests/fast/dom/node-iterator-with-doctype-root-expected.txt

CONSOLE WARNING: 'NodeIterator.detach' is now a no-op, as per DOM (http://dom.spec.whatwg.org/#dom-nodeiterator-detach).
NodeIterator rooted at a DocumentType node not yet associated with a document:
PASS iter.referenceNode.ownerDocument is document
PASS iter.nextNode() is dt
PASS iter.nextNode() is null
PASS iter.previousNode() is dt
PASS iter.previousNode() is null

NodeIterator rooted at a DocumentType node that becomes used by a document after the iterator has been created:
PASS iter.nextNode() is dt
PASS iter.previousNode() is dt
..and after removing it from the document:
PASS iter.nextNode() is dt
PASS iter.previousNode() is dt
PASS successfullyParsed is true

TEST COMPLETE
